# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NemoReconstruct API",
    "description": "Minimal MOV-to-fVDB reconstruction workflow for Isaac Sim and agent orchestration.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reconstructions/upload": {
      "post": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Upload Reconstruction",
        "operationId": "upload_reconstruction_api_v1_reconstructions_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_reconstruction_api_v1_reconstructions_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reconstructions/{reconstruction_id}": {
      "get": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Get Reconstruction",
        "operationId": "get_reconstruction_api_v1_reconstructions__reconstruction_id__get",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconstructionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Delete Reconstruction",
        "operationId": "delete_reconstruction_api_v1_reconstructions__reconstruction_id__delete",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/status": {
      "get": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Get Reconstruction Status",
        "operationId": "get_reconstruction_status_api_v1_reconstructions__reconstruction_id__status_get",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconstructionStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/artifacts": {
      "get": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Get Reconstruction Artifacts",
        "operationId": "get_reconstruction_artifacts_api_v1_reconstructions__reconstruction_id__artifacts_get",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconstructionArtifacts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/download/{artifact}": {
      "get": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Download Artifact",
        "operationId": "download_artifact_api_v1_reconstructions__reconstruction_id__download__artifact__get",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          },
          {
            "name": "artifact",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reconstructions/{reconstruction_id}/retry": {
      "post": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Retry Reconstruction",
        "operationId": "retry_reconstruction_api_v1_reconstructions__reconstruction_id__retry_post",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RetryRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/notes": {
      "patch": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Update Notes",
        "operationId": "update_notes_api_v1_reconstructions__reconstruction_id__notes_patch",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconstructionDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/iterations": {
      "get": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Get Iteration History",
        "operationId": "get_iteration_history_api_v1_reconstructions__reconstruction_id__iterations_get",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IterationHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/iterations/{iteration}/download/splat_ply": {
      "get": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Download Iteration Ply",
        "operationId": "download_iteration_ply_api_v1_reconstructions__reconstruction_id__iterations__iteration__download_splat_ply_get",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          },
          {
            "name": "iteration",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Iteration"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reconstructions/{reconstruction_id}/iterations/{iteration}/verdict": {
      "patch": {
        "tags": [
          "reconstructions"
        ],
        "summary": "Update Iteration Verdict",
        "operationId": "update_iteration_verdict_api_v1_reconstructions__reconstruction_id__iterations__iteration__verdict_patch",
        "parameters": [
          {
            "name": "reconstruction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Reconstruction Id"
            }
          },
          {
            "name": "iteration",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Iteration"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Iteration Verdict Api V1 Reconstructions  Reconstruction Id  Iterations  Iteration  Verdict Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_reconstruction_api_v1_reconstructions_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "frame_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Rate"
          },
          "sequential_matcher_overlap": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequential Matcher Overlap"
          },
          "colmap_mapper_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colmap Mapper Type"
          },
          "colmap_max_num_features": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colmap Max Num Features"
          },
          "fvdb_max_epochs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fvdb Max Epochs"
          },
          "fvdb_sh_degree": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fvdb Sh Degree"
          },
          "fvdb_image_downsample_factor": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fvdb Image Downsample Factor"
          },
          "splat_only_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Splat Only Mode"
          }
        },
        "type": "object",
        "required": [
          "file",
          "name"
        ],
        "title": "Body_upload_reconstruction_api_v1_reconstructions_upload_post"
      },
      "DeleteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DeleteResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}